+++ docs/source/utils.rst
@@ -5,4 +5,6 @@ torchgeometry.utils
 
 .. autofunction:: tensor_to_image
 .. autofunction:: image_to_tensor
+.. autofunction:: create_meshgrid
+.. autofunction:: normalize_pixel_coordinates
