+++ test/test_pinhole.py
@@ -145,7 +145,9 @@ class TestPinholeCamera:
         assert torch.allclose(pinhole_1.ty, pinhole_11.ty)
         assert torch.allclose(pinhole_1.tz, pinhole_11.tz)
         assert torch.allclose(pinhole_1.rt_matrix, pinhole_11.rt_matrix)
-        assert torch.allclose(pinhole_1.camera_matrix, pinhole_11.camera_matrix)
+        assert torch.allclose(
+            pinhole_1.camera_matrix,
+            pinhole_11.camera_matrix)
         assert torch.allclose(
             pinhole_1.rotation_matrix, pinhole_11.rotation_matrix)
         assert torch.allclose(
@@ -167,14 +169,18 @@ class TestPinholeCamera:
         pinhole = tgm.PinholeCamera(intrinsics, extrinsics, height, width)
         pinhole_scale = pinhole.scale(scale_factor)
 
-        assert utils.check_equal_torch(pinhole_scale.intrinsics[..., 0, 0],
-                                       pinhole.intrinsics[..., 0, 0] * scale_val)  # fx
-        assert utils.check_equal_torch(pinhole_scale.intrinsics[..., 1, 1],
-                                       pinhole.intrinsics[..., 1, 1] * scale_val)  # fy
-        assert utils.check_equal_torch(pinhole_scale.intrinsics[..., 0, 2],
-                                       pinhole.intrinsics[..., 0, 2] * scale_val)  # cx
-        assert utils.check_equal_torch(pinhole_scale.intrinsics[..., 1, 2],
-                                       pinhole.intrinsics[..., 1, 2] * scale_val)  # cy
+        assert utils.check_equal_torch(
+            pinhole_scale.intrinsics[..., 0, 0],
+            pinhole.intrinsics[..., 0, 0] * scale_val)  # fx
+        assert utils.check_equal_torch(
+            pinhole_scale.intrinsics[..., 1, 1],
+            pinhole.intrinsics[..., 1, 1] * scale_val)  # fy
+        assert utils.check_equal_torch(
+            pinhole_scale.intrinsics[..., 0, 2],
+            pinhole.intrinsics[..., 0, 2] * scale_val)  # cx
+        assert utils.check_equal_torch(
+            pinhole_scale.intrinsics[..., 1, 2],
+            pinhole.intrinsics[..., 1, 2] * scale_val)  # cy
         assert utils.check_equal_torch(
             pinhole_scale.height,
             pinhole.height * scale_val)
@@ -183,50 +189,55 @@ class TestPinholeCamera:
             pinhole.width * scale_val)
 
     def test_pinhole_camera_scale_inplace(self):
-        batch_size=2
-        height, width=4, 6
-        fx, fy, cx, cy=1, 2, width / 2, height / 2
-        tx, ty, tz=1, 2, 3
-        scale_val=2.0
+        batch_size = 2
+        height, width = 4, 6
+        fx, fy, cx, cy = 1, 2, width / 2, height / 2
+        tx, ty, tz = 1, 2, 3
+        scale_val = 2.0
 
-        intrinsics=self._create_intrinsics(batch_size, fx, fy, cx, cy)
-        extrinsics=self._create_extrinsics(batch_size, tx, ty, tz)
-        height=torch.ones(batch_size) * height
-        width=torch.ones(batch_size) * width
-        scale_factor=torch.ones(batch_size) * scale_val
+        intrinsics = self._create_intrinsics(batch_size, fx, fy, cx, cy)
+        extrinsics = self._create_extrinsics(batch_size, tx, ty, tz)
+        height = torch.ones(batch_size) * height
+        width = torch.ones(batch_size) * width
+        scale_factor = torch.ones(batch_size) * scale_val
 
-        pinhole=tgm.PinholeCamera(intrinsics, extrinsics, height, width)
-        pinhole_scale=pinhole.clone()
+        pinhole = tgm.PinholeCamera(intrinsics, extrinsics, height, width)
+        pinhole_scale = pinhole.clone()
         pinhole_scale.scale_(scale_factor)
 
-        assert utils.check_equal_torch(pinhole_scale.intrinsics[..., 0, 0],
-                                       pinhole.intrinsics[..., 0, 0] * scale_val)  # fx
-        assert utils.check_equal_torch(pinhole_scale.intrinsics[..., 1, 1],
-                                       pinhole.intrinsics[..., 1, 1] * scale_val)  # fy
-        assert utils.check_equal_torch(pinhole_scale.intrinsics[..., 0, 2],
-                                       pinhole.intrinsics[..., 0, 2] * scale_val)  # cx
-        assert utils.check_equal_torch(pinhole_scale.intrinsics[..., 1, 2],
-                                       pinhole.intrinsics[..., 1, 2] * scale_val)  # cy
+        assert utils.check_equal_torch(
+            pinhole_scale.intrinsics[..., 0, 0],
+            pinhole.intrinsics[..., 0, 0] * scale_val)  # fx
+        assert utils.check_equal_torch(
+            pinhole_scale.intrinsics[..., 1, 1],
+            pinhole.intrinsics[..., 1, 1] * scale_val)  # fy
+        assert utils.check_equal_torch(
+            pinhole_scale.intrinsics[..., 0, 2],
+            pinhole.intrinsics[..., 0, 2] * scale_val)  # cx
+        assert utils.check_equal_torch(
+            pinhole_scale.intrinsics[..., 1, 2],
+            pinhole.intrinsics[..., 1, 2] * scale_val)  # cy
         assert utils.check_equal_torch(
             pinhole_scale.height, pinhole.height * scale_val)
         assert utils.check_equal_torch(
             pinhole_scale.width, pinhole.width * scale_val)
 
+
 @pytest.mark.parametrize("device_type", TEST_DEVICES)
 @pytest.mark.parametrize("batch_size", [1, 2, 5, 6])
 def test_scale_pinhole(batch_size, device_type):
     # generate input data
-    device=torch.device(device_type)
-    pinholes=torch.rand(batch_size, 12).to(device)
-    scales=torch.rand(batch_size).to(device)
+    device = torch.device(device_type)
+    pinholes = torch.rand(batch_size, 12).to(device)
+    scales = torch.rand(batch_size).to(device)
 
-    pinholes_scale=tgm.scale_pinhole(pinholes, scales)
+    pinholes_scale = tgm.scale_pinhole(pinholes, scales)
     assert utils.check_equal_torch(
         pinholes_scale[..., :6] / scales.unsqueeze(-1), pinholes[..., :6])
 
     # evaluate function gradient
-    pinholes=utils.tensor_to_gradcheck_var(pinholes)  # to var
-    scales=utils.tensor_to_gradcheck_var(scales)  # to var
+    pinholes = utils.tensor_to_gradcheck_var(pinholes)  # to var
+    scales = utils.tensor_to_gradcheck_var(scales)  # to var
     assert gradcheck(tgm.scale_pinhole, (pinholes, scales,),
                      raise_exception=True)
 
@@ -235,21 +246,21 @@ def test_scale_pinhole(batch_size, device_type):
 @pytest.mark.parametrize("batch_size", [1, 2, 5, 6])
 def test_pinhole_matrix(batch_size, device_type):
     # generate input data
-    image_height, image_width=32., 32.
-    cx, cy=image_width / 2, image_height / 2
-    fx, fy=1., 1.
-    rx, ry, rz=0., 0., 0.
-    tx, ty, tz=0., 0., 0.
-    offset_x=10.  # we will apply a 10units offset to `i` camera
-    eps=1e-6
+    image_height, image_width = 32., 32.
+    cx, cy = image_width / 2, image_height / 2
+    fx, fy = 1., 1.
+    rx, ry, rz = 0., 0., 0.
+    tx, ty, tz = 0., 0., 0.
+    offset_x = 10.  # we will apply a 10units offset to `i` camera
+    eps = 1e-6
 
-    pinhole=utils.create_pinhole(
+    pinhole = utils.create_pinhole(
         fx, fy, cx, cy, image_height, image_width, rx, ry, rx, tx, ty, tz)
-    pinhole=pinhole.repeat(batch_size, 1).to(torch.device(device_type))
+    pinhole = pinhole.repeat(batch_size, 1).to(torch.device(device_type))
 
-    pinhole_matrix=tgm.pinhole_matrix(pinhole)
+    pinhole_matrix = tgm.pinhole_matrix(pinhole)
 
-    ones=torch.ones(batch_size)
+    ones = torch.ones(batch_size)
     assert bool((pinhole_matrix[:, 0, 0] == fx * ones).all())
     assert bool((pinhole_matrix[:, 1, 1] == fy * ones).all())
     assert bool((pinhole_matrix[:, 0, 2] == cx * ones).all())
@@ -259,7 +270,7 @@ def test_pinhole_matrix(batch_size, device_type):
     assert tgm.PinholeMatrix()(pinhole).shape == (batch_size, 4, 4)
 
     # evaluate function gradient
-    pinhole=utils.tensor_to_gradcheck_var(pinhole)  # to var
+    pinhole = utils.tensor_to_gradcheck_var(pinhole)  # to var
     assert gradcheck(tgm.pinhole_matrix, (pinhole,),
                      raise_exception=True)
 
@@ -268,21 +279,21 @@ def test_pinhole_matrix(batch_size, device_type):
 @pytest.mark.parametrize("batch_size", [1, 2, 5, 6])
 def test_inverse_pinhole_matrix(batch_size, device_type):
     # generate input data
-    image_height, image_width=32., 32.
-    cx, cy=image_width / 2, image_height / 2
-    fx, fy=1., 1.
-    rx, ry, rz=0., 0., 0.
-    tx, ty, tz=0., 0., 0.
-    offset_x=10.  # we will apply a 10units offset to `i` camera
-    eps=1e-6
+    image_height, image_width = 32., 32.
+    cx, cy = image_width / 2, image_height / 2
+    fx, fy = 1., 1.
+    rx, ry, rz = 0., 0., 0.
+    tx, ty, tz = 0., 0., 0.
+    offset_x = 10.  # we will apply a 10units offset to `i` camera
+    eps = 1e-6
 
-    pinhole=utils.create_pinhole(
+    pinhole = utils.create_pinhole(
         fx, fy, cx, cy, image_height, image_width, rx, ry, rx, tx, ty, tz)
-    pinhole=pinhole.repeat(batch_size, 1).to(torch.device(device_type))
+    pinhole = pinhole.repeat(batch_size, 1).to(torch.device(device_type))
 
-    pinhole_matrix=tgm.inverse_pinhole_matrix(pinhole)
+    pinhole_matrix = tgm.inverse_pinhole_matrix(pinhole)
 
-    ones=torch.ones(batch_size)
+    ones = torch.ones(batch_size)
     assert utils.check_equal_torch(pinhole_matrix[:, 0, 0], (1. / fx) * ones)
     assert utils.check_equal_torch(pinhole_matrix[:, 1, 1], (1. / fy) * ones)
     assert utils.check_equal_torch(
@@ -294,7 +305,7 @@ def test_inverse_pinhole_matrix(batch_size, device_type):
     assert tgm.InversePinholeMatrix()(pinhole).shape == (batch_size, 4, 4)
 
     # evaluate function gradient
-    pinhole=utils.tensor_to_gradcheck_var(pinhole)  # to var
+    pinhole = utils.tensor_to_gradcheck_var(pinhole)  # to var
     assert gradcheck(tgm.pinhole_matrix, (pinhole,),
                      raise_exception=True)
 
@@ -303,20 +314,20 @@ def test_inverse_pinhole_matrix(batch_size, device_type):
 @pytest.mark.parametrize("batch_size", [1, 2, 5, 6])
 def test_homography_i_H_ref(batch_size, device_type):
     # generate input data
-    device=torch.device(device_type)
-    image_height, image_width=32., 32.
-    cx, cy=image_width / 2, image_height / 2
-    fx, fy=1., 1.
-    rx, ry, rz=0., 0., 0.
-    tx, ty, tz=0., 0., 0.
-    offset_x=10.  # we will apply a 10units offset to `i` camera
-    eps=1e-6
+    device = torch.device(device_type)
+    image_height, image_width = 32., 32.
+    cx, cy = image_width / 2, image_height / 2
+    fx, fy = 1., 1.
+    rx, ry, rz = 0., 0., 0.
+    tx, ty, tz = 0., 0., 0.
+    offset_x = 10.  # we will apply a 10units offset to `i` camera
+    eps = 1e-6
 
-    pinhole_ref=utils.create_pinhole(
+    pinhole_ref = utils.create_pinhole(
         fx, fy, cx, cy, image_height, image_width, rx, ry, rx, tx, ty, tz)
-    pinhole_ref=pinhole_ref.repeat(batch_size, 1).to(device)
+    pinhole_ref = pinhole_ref.repeat(batch_size, 1).to(device)
 
-    pinhole_i=utils.create_pinhole(
+    pinhole_i = utils.create_pinhole(
         fx,
         fy,
         cx,
@@ -329,14 +340,14 @@ def test_homography_i_H_ref(batch_size, device_type):
         tx + offset_x,
         ty,
         tz)
-    pinhole_i=pinhole_i.repeat(batch_size, 1).to(device)
+    pinhole_i = pinhole_i.repeat(batch_size, 1).to(device)
 
     # compute homography from ref to i
-    i_H_ref=tgm.homography_i_H_ref(pinhole_i, pinhole_ref) + eps
-    i_H_ref_inv=torch.inverse(i_H_ref)
+    i_H_ref = tgm.homography_i_H_ref(pinhole_i, pinhole_ref) + eps
+    i_H_ref_inv = torch.inverse(i_H_ref)
 
     # compute homography from i to ref
-    ref_H_i=tgm.homography_i_H_ref(pinhole_ref, pinhole_i) + eps
+    ref_H_i = tgm.homography_i_H_ref(pinhole_ref, pinhole_i) + eps
     assert utils.check_equal_torch(i_H_ref_inv, ref_H_i)
 
     # evaluate function gradient
