+++ torchgeometry/core/pinhole.py
@@ -1,11 +1,11 @@
-from typing import Iterable
+from typing import Iterable, Optional
 import warnings
 
 import torch
 import torch.nn as nn
 
-from .conversions import rtvec_to_pose
-from .transformations import inverse_pose
+from torchgeometry.core.transformations import inverse_pose
+from torchgeometry.core.conversions import rtvec_to_pose, transform_points
 
 
 __all__ = [
@@ -16,6 +16,8 @@ __all__ = [
     "inverse_pinhole_matrix",
     "scale_pinhole",
     "homography_i_H_ref",
+    "pixel2cam",
+    "cam2pixel",
     # layer api
     "PinholeMatrix",
     "InversePinholeMatrix",
@@ -534,6 +536,78 @@ def homography_i_H_ref(pinhole_i, pinhole_ref):
     return torch.matmul(
         pinhole_matrix(pinhole_i),
         torch.matmul(i_pose_ref, inverse_pinhole_matrix(pinhole_ref)))
+
+# based on:
+# https://github.com/ClementPinard/SfmLearner-Pytorch/blob/master/inverse_warp.py#L26
+
+
+def pixel2cam(depth: torch.Tensor, intrinsics_inv: torch.Tensor,
+              pixel_coords: torch.Tensor) -> torch.Tensor:
+    r"""Transform coordinates in the pixel frame to the camera frame.
+
+    Args:
+        depth (torch.Tensor): the source depth maps. Shape must be Bx1xHxW.
+        intrinsics_inv (torch.Tensor): the inverse intrinsics camera matrix.
+          Shape must be Bx4x4.
+        pixel_coords (torch.Tensor): the grid with the homogeneous camera
+          coordinates. Shape must be BxHxWx3.
+
+    Returns:
+        torch.Tensor: array of (u, v, 1) cam coordinates with shape BxHxWx3.
+    """
+    if not len(depth.shape) == 4 and depth.shape[1] == 1:
+        raise ValueError("Input depth has to be in the shape of "
+                         "Bx1xHxW. Got {}".format(depth.shape))
+    if not len(intrinsics_inv.shape) == 3:
+        raise ValueError("Input intrinsics_inv has to be in the shape of "
+                         "Bx4x4. Got {}".format(intrinsics_inv.shape))
+    if not len(pixel_coords.shape) == 4 and pixel_coords.shape[3] == 3:
+        raise ValueError("Input pixel_coords has to be in the shape of "
+                         "BxHxWx3. Got {}".format(intrinsics_inv.shape))
+    cam_coords: torch.Tensor = transform_points(
+        intrinsics_inv[:, None], pixel_coords)
+    return cam_coords * depth.permute(0, 2, 3, 1)
+
+
+# based on
+# https://github.com/ClementPinard/SfmLearner-Pytorch/blob/master/inverse_warp.py#L43
+
+def cam2pixel(
+        cam_coords_src: torch.Tensor,
+        dst_proj_src: torch.Tensor,
+        eps: Optional[float] = 1e-6) -> torch.Tensor:
+    r"""Transform coordinates in the camera frame to the pixel frame.
+
+    Args:
+        cam_coords (torch.Tensor): pixel coordinates defined in the first
+          camera coordinates system. Shape must be BxHxWx3.
+        dst_proj_src (torch.Tensor): the projection matrix between the
+          reference and the non reference camera frame. Shape must be Bx4x4.
+
+    Returns:
+        torch.Tensor: array of [-1, 1] coordinates of shape BxHxWx2.
+    """
+    if not len(cam_coords_src.shape) == 4 and cam_coords_src.shape[3] == 3:
+        raise ValueError("Input cam_coords_src has to be in the shape of "
+                         "BxHxWx3. Got {}".format(cam_coords_src.shape))
+    if not len(dst_proj_src.shape) == 3 and dst_proj_src.shape[-2:] == (4, 4):
+        raise ValueError("Input dst_proj_src has to be in the shape of "
+                         "Bx4x4. Got {}".format(dst_proj_src.shape))
+    b, h, w, _ = cam_coords_src.shape
+    # apply projection matrix to points
+    point_coords: torch.Tensor = transform_points(
+        dst_proj_src[:, None], cam_coords_src)
+    x_coord: torch.Tensor = point_coords[..., 0]
+    y_coord: torch.Tensor = point_coords[..., 1]
+    z_coord: torch.Tensor = point_coords[..., 2]
+
+    # compute pixel coordinates
+    u_coord: torch.Tensor = x_coord / (z_coord + eps)
+    v_coord: torch.Tensor = y_coord / (z_coord + eps)
+
+    # stack and return the coordinates, that's the actual flow
+    pixel_coords_dst: torch.Tensor = torch.stack([u_coord, v_coord], dim=-1)
+    return pixel_coords_dst  # (B*N)xHxWx2
 
 
 # layer api
